Extract queries to gql files
