+++ manifest.json
@@ -24,7 +24,8 @@
     "vtex.styleguide": "9.x",
     "vtex.store-icons": "0.x",
     "vtex.pixel-manager": "1.x",
-    "vtex.react-portal": "0.x"
+    "vtex.react-portal": "0.x",
+    "vtex.store-graphql": "2.x"
   },
   "$schema": "https://raw.githubusercontent.com/vtex/node-vtex-api/master/gen/manifest.schema"
 }
