+++ react/localState/graphql/fullMinicartQuery.gql
@@ -1,0 +1,7 @@
+query {
+  minicart @client {
+    items
+    orderForm
+    isOpen
+  }
+}
