+++ react/localState/graphql/minicartIsOpenQuery.gql
@@ -1,0 +1,5 @@
+query {
+  minicart @client {
+    isOpen
+  }
+}
