+++ react/localState/graphql/minicartItemsQuery.gql
@@ -1,0 +1,5 @@
+query {
+  minicart @client {
+    items
+  }
+}
