+++ react/localState/graphql/updateItemsMutation.gql
@@ -1,0 +1,3 @@
+mutation updateItems($items: [MinicartItem]) {
+  updateItems(items: $items) @client
+}
