+++ react/localState/graphql/updateLocalItemStatusMutation.gql
@@ -1,0 +1,3 @@
+mutation updateLocalItemStatus {
+  updateLocalItemStatus @client
+}
