+++ react/localState/graphql/updateOrderFormMutation.gql
@@ -1,0 +1,3 @@
+mutation updateOrderForm($orderForm: [OrderForm]) {
+  updateOrderForm(orderForm: $orderForm) @client
+}
